Create Go-to-top button as own component
Refactor code for go to top arrow button into its own component and CSS
module
