+++ frontend/src/styles/buttons/ScrollToTopButton.module.css
@@ -1,0 +1,23 @@
+.ScrollToTopButton {
+    position: fixed;
+    display: inline-block;
+    right: 2%;
+    bottom: 20%;
+    z-index: 99;
+    cursor: pointer;
+    align-items: center;
+    justify-content: center;
+    object-fit: fill;
+  }
+  
+  .ScrollToTopButton .ScrollToTopButtonHover {
+    display: none;
+    left: 0;
+    position: absolute;
+    object-fit: fill;
+  }
+  
+  .ScrollToTopButton:hover .ScrollToTopButtonHover {
+    display: inline;
+  }
+
